deploy/requirements.txt
asgiref==3.6.0
black==22.12.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
Django==4.1.5
django-cors-headers==3.13.0
django-embed-video==1.4.8
django-extensions==3.2.1
django-resized==1.0.2
django-taggit==3.1.0
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
idna==3.4
MarkupSafe==2.1.1
mypy-extensions==0.4.3
pathspec==0.11.0
Pillow==9.4.0
platformdirs==2.6.2
pytz==2022.7
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
urllib3==1.26.13
Werkzeug==2.2.2
whitenoise==6.3.0
